# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Processing System API
  description: |
    API for payment flows (Authorize.Net sandbox integration). Contains endpoints for Purchase, Authorize, Capture, Cancel (void), Refund, Recurring Billing (Subscriptions), and Webhook handling. 
    JWT bearer is used to protect payment endpoints. Webhooks use HMAC signature validation. 
    All mutating endpoints support Idempotency-Key header for safe retries. 
    All responses include X-Correlation-ID for distributed tracing.
  version: 2.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /payments/purchase:
    post:
      summary: Purchase (authorize + capture)
      tags:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '201':
          description: Purchase created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '422':
          description: Payment declined by gateway (card declined, insufficient funds, fraud hold)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDeclinedResponse'
        '502':
          description: Gateway error (Authorize.Net returned an error or is unreachable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayErrorResponse'
        '504':
          description: Gateway timeout (Authorize.Net did not respond in time)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTimeoutResponse'
        '409':
          description: Idempotency conflict (different request body for same key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
  /payments/authorize:
    post:
      summary: Authorize only (two-step)
      tags:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '201':
          description: Authorization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '422':
          description: Payment declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDeclinedResponse'
        '502':
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayErrorResponse'
        '504':
          description: Gateway timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTimeoutResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/capture:
    post:
      summary: Capture a previously authorized transaction
      tags:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '200':
          description: Capture result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Capture declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDeclinedResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '502':
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayErrorResponse'
        '504':
          description: Gateway timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTimeoutResponse'
  /payments/cancel:
    post:
      summary: Void (cancel) a previously authorized transaction
      tags:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: Void result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Cancel declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDeclinedResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '502':
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayErrorResponse'
        '504':
          description: Gateway timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTimeoutResponse'
  /payments/refund:
    post:
      summary: Refund (full or partial) a captured transaction
      tags:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Original transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Refund declined by gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDeclinedResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
        '502':
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayErrorResponse'
        '504':
          description: Gateway timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayTimeoutResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Card:
      type: object
      properties:
        number:
          type: string
          description: Card PAN
          example: "4111111111111111"
        expMonth:
          type: integer
          example: 12
        expYear:
          type: integer
          example: 2030
        cvv:
          type: string
          example: "123"
      required:
        - number
        - expMonth
        - expYear
        - cvv
    PurchaseRequest:
      type: object
      properties:
        amount:
          type: number
          format: decimal
          example: 12.34
        currency:
          type: string
          example: USD
        card:
          $ref: '#/components/schemas/Card'
        orderId:
          type: string
          example: ext-100
      required:
        - amount
        - card
    PurchaseResponse:
      type: object
      properties:
        order_id:
          type: integer
          example: 100
        transaction_id:
          type: string
          example: prov-123
        status:
          type: string
          example: success
    CaptureRequest:
      type: object
      properties:
        transactionId:
          type: string
          example: prov-auth-123
        amount:
          type: number
          format: decimal
          example: 10.00
      required:
        - transactionId
    CaptureResponse:
      type: object
      properties:
        transaction_id:
          type: string
          example: prov-capture-123
        status:
          type: string
          example: success
    CancelRequest:
      type: object
      properties:
        transactionId:
          type: string
          example: prov-auth-123
      required:
        - transactionId
    RefundRequest:
      type: object
      properties:
        transactionId:
          type: string
          example: prov-captured-123
        amount:
          type: number
          format: decimal
          example: 5.00
        last4:
          type: string
          description: Last 4 digits of the card (required by Authorize.Net refunds)
          example: "1111"
      required:
        - transactionId
        - last4
    RefundResponse:
      type: object
      properties:
        refund_transaction_id:
          type: string
          example: prov-refund-123
        status:
          type: string
          example: success
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "transaction not found"
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            amount: "must be >= 0.01"
            card.number: "invalid card number"
...
    PaymentDeclinedResponse:
      type: object
      description: Returned when Authorize.Net declines the transaction (card declined, AVS mismatch, fraud hold)
      properties:
        error_code:
          type: string
          enum: [CARD_DECLINED, INSUFFICIENT_FUNDS, FRAUD_SUSPECTED, AVS_MISMATCH]
          example: CARD_DECLINED
        message:
          type: string
          example: "Transaction was declined by the issuing bank"
        category:
          type: string
          enum: [DECLINE]
          example: DECLINE
        retryable:
          type: boolean
          example: false
        request_id:
          type: string
          example: "req_abc12345"
    GatewayErrorResponse:
      type: object
      description: Returned when Authorize.Net returns an error or is unreachable
      properties:
        error_code:
          type: string
          enum: [GATEWAY_ERROR, GATEWAY_UNAVAILABLE, GATEWAY_COMMUNICATION_ERROR]
          example: GATEWAY_ERROR
        message:
          type: string
          example: "Payment gateway returned an error"
        category:
          type: string
          enum: [GATEWAY]
          example: GATEWAY
        retryable:
          type: boolean
          example: true
        retry_after_seconds:
          type: integer
          description: Suggested wait time before retry (exponential backoff)
          example: 5
        request_id:
          type: string
          example: "req_abc12345"
    GatewayTimeoutResponse:
      type: object
      description: Returned when Authorize.Net does not respond within timeout window
      properties:
        error_code:
          type: string
          enum: [GATEWAY_TIMEOUT]
          example: GATEWAY_TIMEOUT
        message:
          type: string
          example: "Payment gateway did not respond in time"
        category:
          type: string
          enum: [GATEWAY]
          example: GATEWAY
        retryable:
          type: boolean
          example: true
        retry_after_seconds:
          type: integer
          example: 10
        request_id:
          type: string
          example: "req_abc12345"
    RateLimitResponse:
      type: object
      description: Returned when API rate limit is exceeded (Resilience4j @RateLimiter)
      properties:
        error_code:
          type: string
          enum: [RATE_LIMIT_EXCEEDED]
          example: RATE_LIMIT_EXCEEDED
        message:
          type: string
          example: "Too many requests. Please retry after a short delay."
        retryable:
          type: boolean
          example: true
        retry_after_seconds:
          type: integer
          description: Suggested wait time before retrying
          example: 1
        suggestion:
          type: string
          example: "Reduce request frequency or implement backoff"
        request_id:
          type: string
          example: "req_abc12345"
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Idempotency key for safe retries (24h TTL). Same key + same body returns cached response.
      schema:
        type: string
        example: "purchase-order-123-v1"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
      schema:
        type: string
        example: "abc12345-6789-def0"
...